requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
PyYAML>=6.0
aiosqlite>=0.19.0
pytz>=2024.1

# Optional: free alternative to Polygon.io (used by scanner/yfinance_client.py)
# yfinance>=0.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
